djchat/requirements.txt
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
astroid==2.4.2
attrs==18.2.0
autopep8==1.5.5
awscli==1.31.13
awsume==4.5.4
backports.zoneinfo==0.2.1
boto3==1.33.13
botocore==1.33.13
cached-property==1.5.2
certifi==2020.12.5
channels==4.0.0
charset-normalizer==2.0.4
click==8.1.7
click-completion==0.5.2
click-didyoumean==0.0.3
click-help-colors==0.9.1
colorama==0.4.3
distlib==0.3.1
Django==3.2.25
django-cors-headers==4.1.0
djangorestframework==3.15.1
docutils==0.16
drf-spectacular==0.27.2
exceptiongroup==1.2.1
filelock==3.0.12
flake8==3.8.4
future==0.18.2
gql==3.0.0a6
gradient==2.0.6
gradient-utils==0.3.2
graphql-core==3.1.7
h11==0.14.0
h5py==3.4.0
halo==0.0.31
httptools==0.6.0
hyperopt==0.1.2
idna==3.2
importlib-metadata==3.4.0
importlib-resources==5.12.0
inflection==0.5.1
isort==5.7.0
jedi==0.17.2
Jinja2==3.0.1
jmespath==1.0.1
jsonschema==4.17.3
lazy-object-proxy==1.4.3
log-symbols==0.0.14
MarkupSafe==2.0.1
marshmallow==2.21.0
mccabe==0.6.1
multidict==6.0.4
networkx==2.6.2
numpy==1.18.5
parso==0.7.1
Pillow==9.5.0
pipenv==2020.11.15
pkgutil_resolve_name==1.3.10
pluggy==0.13.1
progressbar2==3.53.1
prometheus-client==0.9.0
psutil==5.9.8
pyasn1==0.5.1
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
pylint==2.6.0
pymongo==3.12.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-utils==2.5.6
pytz==2024.1
PyYAML==5.4.1
requests==2.26.0
requests-toolbelt==0.9.1
rope==0.18.0
rsa==4.7.2
s3transfer==0.8.2
scipy==1.7.1
shellingham==1.4.0
six==1.15.0
sniffio==1.3.1
snowballstemmer==2.1.0
spinners==0.0.24
sqlparse==0.4.4
termcolor==1.1.0
terminaltables==3.1.0
toml==0.10.2
tqdm==4.62.2
typed-ast==1.4.2
typing_extensions==4.7.1
ujson==4.0.2
uritemplate==4.1.1
urllib3==1.26.6
uvicorn==0.22.0
uvloop==0.18.0
virtualenv==20.4.2
virtualenv-clone==0.5.4
watchfiles==0.20.0
websocket-client==0.57.0
websockets==11.0.3
wrapt==1.12.1
yapf==0.30.0
yarl==1.9.3
zipp==3.4.0
